♻️ Découpage propre du projector
